+++ windows/keyboard_shortcut_mapping_plugin.cpp
@@ -1,0 +1,58 @@
+#include "keyboard_shortcut_mapping_plugin.h"
+
+// This must be included before many other Windows headers.
+#include <windows.h>
+
+#include <VersionHelpers.h>
+
+#include <flutter/method_channel.h>
+#include <flutter/plugin_registrar_windows.h>
+#include <flutter/standard_method_codec.h>
+
+#include <memory>
+#include <sstream>
+
+namespace keyboard_shortcut_mapping {
+
+// static
+void KeyboardShortcutMappingPlugin::RegisterWithRegistrar(
+    flutter::PluginRegistrarWindows *registrar) {
+  auto channel =
+      std::make_unique<flutter::MethodChannel<flutter::EncodableValue>>(
+          registrar->messenger(), "keyboard_shortcut_mapping",
+          &flutter::StandardMethodCodec::GetInstance());
+
+  auto plugin = std::make_unique<KeyboardShortcutMappingPlugin>();
+
+  channel->SetMethodCallHandler(
+      [plugin_pointer = plugin.get()](const auto &call, auto result) {
+        plugin_pointer->HandleMethodCall(call, std::move(result));
+      });
+
+  registrar->AddPlugin(std::move(plugin));
+}
+
+KeyboardShortcutMappingPlugin::KeyboardShortcutMappingPlugin() {}
+
+KeyboardShortcutMappingPlugin::~KeyboardShortcutMappingPlugin() {}
+
+void KeyboardShortcutMappingPlugin::HandleMethodCall(
+    const flutter::MethodCall<flutter::EncodableValue> &method_call,
+    std::unique_ptr<flutter::MethodResult<flutter::EncodableValue>> result) {
+  if (method_call.method_name().compare("getPlatformVersion") == 0) {
+    std::ostringstream version_stream;
+    version_stream << "Windows ";
+    if (IsWindows10OrGreater()) {
+      version_stream << "10+";
+    } else if (IsWindows8OrGreater()) {
+      version_stream << "8";
+    } else if (IsWindows7OrGreater()) {
+      version_stream << "7";
+    }
+    result->Success(flutter::EncodableValue(version_stream.str()));
+  } else {
+    result->NotImplemented();
+  }
+}
+
+}  // namespace keyboard_shortcut_mapping