+++ windows/test/keyboard_shortcut_mapping_plugin_test.cpp
@@ -1,0 +1,39 @@
+#include <flutter/method_call.h>
+#include <flutter/method_result_functions.h>
+#include <flutter/standard_method_codec.h>
+#include <gtest/gtest.h>
+#include <windows.h>
+
+#include <memory>
+#include <string>
+#include <variant>
+
+#include "keyboard_shortcut_mapping_plugin.h"
+
+namespace keyboard_shortcut_mapping {
+namespace test {
+
+namespace {
+
+using flutter::EncodableValue;
+using flutter::MethodCall;
+using flutter::MethodResultFunctions;
+
+}  // namespace
+
+TEST(KeyboardShortcutMappingPlugin, GetPlatformVersion) {
+  KeyboardShortcutMappingPlugin plugin;
+  std::string result_string;
+  plugin.HandleMethodCall(
+      MethodCall("getPlatformVersion", std::make_unique<EncodableValue>()),
+      std::make_unique<MethodResultFunctions<>>(
+          [&result_string](const EncodableValue* result) {
+            result_string = std::get<std::string>(*result);
+          },
+          nullptr, nullptr));
+
+  EXPECT_TRUE(result_string.rfind("Windows ", 0) == 0);
+}
+
+}  // namespace test
+}  // namespace keyboard_shortcut_mapping