+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 * Fix: Shortened package description for pub.dev compliance
 * Fix: Added issue_tracker URL to pubspec.yaml
+* Feat: Add Windows support for keyboard shortcut mapping plugin
 
 ## 0.0.1